Enable keyword redirection for trace
BREAKING CHANGE: Changed `/cr/trace` response format.
BREAKING CHANGE: Changed default behavior for definition search in `/cr/{rule_id}`

If the trace endpoint is called with a keyword rule, it now supports redirection to the actual definition of that rule, same as the rule endpoint. This necessitated the change of the response from an array to a dict, so that we can send the actual rule that was queried.

I also changed the default of the rule endpoint to *not* default to definition redirection in order to be consistent with the rest of the API.