Read proxy headers in container (#80)
The proxy headers are only accepted from the default localhost, which can cause incorrect redirects in certain deployments. The container should always be deployed behind a reverse proxy anyways, so there should be no harm in accepting the headers from anywhere (making it the responsibility of the proxy to set them correctly).