Update rule mention regex
The regex for checking whether a change is just a reference to a renumbered rule handles cases such as "[...] (see rule 104.3a)."

Some of these are not written as a part of the sentence though, and asa such are punctuated differently, i.e. "[...]. (See rule 104.3a.)"

Having the period inside the parentheses isn't supported by the matching regex, so I'm updating it to properly detect such references as well.